Restructure mock repository providers like the data project
