+++ packages/server/testing/mock/entities/api/mock-ability-module.provider.ts
@@ -1,5 +1,5 @@
 import { mockAbilityDB } from '@dnd-mapp/data/testing';
-import { AbilityRepository, AbilityService } from '../../../src/app/entities/ability';
+import { AbilityRepository, AbilityService } from '../../../../src/app/entities/ability';
 
 export const mockAbilityModuleProviders = [
     AbilityService,
