+++ packages/server/testing/mock/entities/api/mock-race-module.provider.ts
@@ -1,5 +1,5 @@
 import { mockRaceDB } from '@dnd-mapp/data/testing';
-import { RaceRepository, RaceService } from '../../../src/app/entities/race';
+import { RaceRepository, RaceService } from '../../../../src/app/entities/race';
 
 export const mockRaceModuleProviders = [
     RaceService,
