+++ packages/server/testing/mock/entities/api/mock-skill-module.provider.ts
@@ -1,5 +1,5 @@
 import { mockSkillDB } from '@dnd-mapp/data/testing';
-import { SkillRepository, SkillService } from '../../../src/app/entities/skill';
+import { SkillRepository, SkillService } from '../../../../src/app/entities/skill';
 
 export const mockSkillModuleProviders = [
     SkillService,
