+++ packages/server/testing/mock/entities/authentication/index.ts
@@ -1,0 +1,2 @@
+export * from './mock-user-module.provider';
+export * from './mock-user-role-module.provider';
