+++ packages/server/testing/mock/entities/authentication/mock-user-module.provider.ts
@@ -1,5 +1,5 @@
 import { mockUserDB } from '@dnd-mapp/data/testing';
-import { UserRepository, UserService } from '../../../src/app/entities/user';
+import { UserRepository, UserService } from '../../../../src/app/entities/user';
 
 export const mockUserModuleProviders = [
     UserService,
