+++ packages/server/testing/mock/entities/authentication/mock-user-role-module.provider.ts
@@ -1,5 +1,5 @@
 import { mockUserRoleDB } from '@dnd-mapp/data/testing';
-import { UserRoleRepository, UserRoleService } from '../../../src/app/entities/user-role';
+import { UserRoleRepository, UserRoleService } from '../../../../src/app/entities/user-role';
 
 export const mockUserRoleModuleProviders = [
     UserRoleService,
