+++ packages/server/testing/mock/entities/index.ts
@@ -1,5 +1,2 @@
-export * from './mock-ability-module.provider';
-export * from './mock-race-module.provider';
-export * from './mock-skill-module.provider';
-export * from './mock-user-module.provider';
-export * from './mock-user-role-module.provider';
+export * from './api';
+export * from './authentication';
